+++ constant.py
@@ -25,7 +25,7 @@ _PASSWORD_FILE = 'password.pickle'  # 密码数据库格式说明：存储一个
 _PASSWORD_EXPORT = '密码导出.txt'
 _Unzip_Temp_Folder = 'UnzipTempFolder'
 _COLOR_ERROR = (254, 67, 101)
-_COLOR_WARNING = (255,215,0)
-_COLOR_SUCCESS = (0,0,0)
-_PATH_7ZIP = './7-Zip/7z.exe'
+_COLOR_WARNING = (255, 215, 0)
+_COLOR_SUCCESS = (0, 0, 0)
+_PATH_7ZIP = '7-Zip/7z.exe'
 _PASSWORD_NONE = 'NonePassword'  # 用于测试是否存在密码的临时密码
