+++ module/function_archive.py
@@ -1,10 +1,14 @@
+# 压缩文件相关方法
+
 import os
 import re
 import subprocess
 from typing import Union
-from constant import _PATH_7ZIP
+
 import filetype
 
+from constant import _PATH_7ZIP
+from module import function_normal
 from module.class_state import State7zResult
 
 
@@ -14,6 +18,7 @@ def find_volume_archives(files_list: list) -> dict:
     :param files_list: list类型，文件列表
     :return: dict类型，{A压缩文件第一个分卷包路径:(A压缩文件全部分卷路径), ...}
     """
+    function_normal.print_function_info()
     # 将文件按所在目录分类
     parent_folder_dict = {}  # {文件父目录:(A文件, B文件), ...}，仅匹配同一文件夹下的分卷压缩文件
     for file in files_list:
@@ -61,6 +66,7 @@ def is_volume_archive(filename: str) -> Union[str, bool]:
     判断传入的文本是否符合分卷压缩文件的规则，是则返回生成第一个分卷包的文件名
     :return: 第一个分卷包的文件名 或 False
     """
+    function_normal.print_function_info()
     pattern_7z = r"^(.+)\.7z\.\d+$"
     pattern_rar = r"^(.+)\.part(\d+)\.rar$"
     pattern_rar_without_suffix = r"^(.+)\.part(\d+)$"
@@ -84,7 +90,7 @@ def is_volume_archive(filename: str) -> Union[str, bool]:
         first_volume_filename = filetitle + f'.part{str(1).zfill(number_type)}'
     # 匹配zip正则（zip分卷文件的第一个包一般都是.zip后缀，所以.zip后缀直接分类为分卷压缩文件）
     elif re.match(pattern_zip, filename):
-        filetitle = re.match(pattern_zip_volume, filename).group(1)
+        filetitle = re.match(pattern_zip, filename).group(1)
         first_volume_filename = filetitle + r'.zip'
     # 匹配zip分卷正则
     elif re.match(pattern_zip_volume, filename):
@@ -107,6 +113,7 @@ def is_archive(filepath: str, use_filetype: bool = True) -> bool:
     :param use_filetype: bool类型，是否使用filetype库进行额外判断
     :return: bool类型，是否为压缩包
     """
+    function_normal.print_function_info()
     if not os.path.exists(filepath):
         return False
     elif os.path.isdir(filepath):
@@ -134,6 +141,7 @@ def is_archive(filepath: str, use_filetype: bool = True) -> bool:
 
 def is_zip_archive(filepath: str) -> bool:
     """单独判断文件是否为zip格式压缩文件"""
+    function_normal.print_function_info()
     archive_type = ['zip']
     kind = filetype.guess(filepath)
     if kind is None:
@@ -148,6 +156,7 @@ def is_zip_archive(filepath: str) -> bool:
 
 def subprocess_run_7z(command_type, file, password):
     """使用run调用7z，仅用于l和t命令"""
+    function_normal.print_function_info()
     command = [_PATH_7ZIP,
                command_type,
                file,
@@ -193,16 +202,18 @@ def subprocess_run_7z(command_type, file, password):
 
     return result
 
+
 def subprocess_7z_l(file, password):
     """调用7z的l指令"""
-
+    function_normal.print_function_info()
     result = subprocess_run_7z('l', file, password)
 
+    return result
 
 
 def subprocess_7z_t(file, password):
     """调用7z的t指令"""
+    function_normal.print_function_info()
     result = subprocess_run_7z('t', file, password)
 
-
-
+    return result
