+++ ui/drop_label.py
@@ -1,8 +1,9 @@
 # 自定义拖入控件
 
 from PySide6.QtCore import Signal
-from PySide6.QtGui import QImage, QPixmap
+from PySide6.QtGui import QPixmap
 from PySide6.QtWidgets import QLabel
+
 from constant import _ICON_DEFAULT, _ICON_DROP
 
 
@@ -14,13 +15,13 @@ class DropLabel(QLabel):
         self.setAcceptDrops(True)
 
         self.icon = _ICON_DEFAULT
+        self.movie_icon = None  # 动图对象，用于显示GIF
         self.last_icon = None
         self.reset_icon(self.icon)
 
-    def reset_icon(self, icon:str):
+    def reset_icon(self, icon: str):
         self.icon = icon
         self.setPixmap(QPixmap(icon))
-
 
     def dragEnterEvent(self, event):
         if event.mimeData().hasUrls():
